make events a separate crate
