+++ Cargo.toml
@@ -15,6 +15,7 @@ wasmtime-wasi = "0.35"
 wasi-common = "0.35"
 wasi-cap-std-sync = "0.35"
 wit-bindgen-wasmtime = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "a79a4be33d76ddf62839ba71602c26a96610ef7c" }
+events = { path = "./crates/events" }
 
 [workspace]
-members = ["demo"]
+members = ["demo", "crates/events"]
